Fix rounding error in StateSpace._unsafe_encode
Numpy arrays use different integer types than Python proper. When they
overflow, they don't covert to larger integer types. Instead they
convert to floats. This makes sense give numpy is written in C, but it's
a problem for us as we encode states of type numpy.ndarray.
Specifically, `encoded += place * x` will change the type of `encoded`
from `long` to `np.float64` if `x` is a numpy integer/float type and
`place * x` overlows. When we get to the return statement of
`StateSpace._unsafe_encode` and convert the result to a `long`, the
value gets rounded.

We fixed this by changing
```
encoded += place * x
```
to
```
encoded += place * long(x)
```
and getting rid of the redundant conversion at the end.

+++ neet/statespace.py
@@ -63,10 +63,10 @@ class StateSpace(object):
         encoded, place = long(0), long(1)
 
         for (x, b) in zip(state, self.shape):
-            encoded += place * x
+            encoded += place * long(x)
             place *= b
 
-        return long(encoded)
+        return encoded
 
     def encode(self, state):
         if state not in self:
